role prop cannot be updated and admin cannot be deleted

+++ @server/@api-user/user.service.ts
@@ -16,9 +16,9 @@ export const getOneUserService = async (paramsId: string) => {
 };
 
 export const deleteOneUserService = async (paramsId: string) => {
-  const query = await User.deleteOne({ _id: paramsId }).exec();
+  const query = await User.deleteOne({ _id: paramsId, role:{$ne: UserRole.Admin} }).exec();
   if (query.deletedCount < 1){
-    notFoundErr('No record found for provided ID to be deleted')
+    notFoundErr('Delete Unsusseccful: Invalid ID or trying to delete the Admin User')
   }
   return query;
 }
@@ -30,6 +30,9 @@ export const updateOneUserPropertyValueService = async (paramsId: string, reques
   }
 
   for (const ops of requestBody) {
+    if (ops.propName === "role") {
+      badRequestErr('<role> property update not allowed')
+    }
     if(!(ops.propName in query)){
       badRequestErr(`invalid property: ${ops.propName}`);
     }
@@ -41,6 +44,10 @@ export const updateOneUserPropertyValueService = async (paramsId: string, reques
 };
 
 export const updateUserPropertyValuesService = async (paramsId: string, requestBody: UserDocument) => {
+  if (requestBody.role) {
+    badRequestErr('<role> property update not allowed')
+  }
+
   const query = await User.findById(paramsId).exec();
   if(!query){
     notFoundErr('No record found for provided ID');
